Improve home page

Closes #60 and #66

See merge request yepopov/team-102-14!23

+++ app/src/main/java/com/example/echonote/data/models/FolderModel.kt
@@ -2,6 +2,10 @@ package com.example.echonote.data.models
 
 import com.example.echonote.data.persistence.IPersistence
 import com.example.echonote.data.entities.Folder
+import com.example.echonote.utils.EmptyArgumentEchoNoteException
+import com.example.echonote.utils.IllegalArgumentEchoNoteException
+import com.example.echonote.utils.IllegalStateEchoNoteException
+import com.example.echonote.utils.NotFoundEchoNoteException
 import kotlinx.datetime.LocalDateTime
 
 class FolderModel(
@@ -25,8 +29,9 @@ class FolderModel(
     }
 
     suspend fun add(title: String, description: String?) {
+        if (title.isEmpty()) throw EmptyArgumentEchoNoteException("Title must not be empty")
         val element = folders.find { it.title == title }
-        if (element != null) throw IllegalArgumentException("Title already in use")
+        if (element != null) throw IllegalArgumentEchoNoteException("Title already in use")
         val currentTime = dateTimeCreator()
         val folder = persistence.createFolder(title, description, currentTime, currentTime)
         folders.add(folder)
@@ -35,25 +40,26 @@ class FolderModel(
 
     suspend fun changeTitle(id: Long, title: String) {
         val element = folders.find { it.id != id && it.title == title }
-        if (element != null) throw IllegalArgumentException("Title $title is already in use")
+        if (element != null) throw IllegalArgumentEchoNoteException("Title $title is already in use")
         val current = folders.find { it.id == id }
-        if(current == null) throw IllegalArgumentException("No id $id exists")
+        if(current == null) throw NotFoundEchoNoteException("No id $id exists")
         current.title = title
         current.updated_on = dateTimeCreator()
-        persistence.saveFolders(folders)
+        persistence.saveFolder(current)
         notifySubscribers()
     }
 
     suspend fun changeDescription(id: Long, description: String?) {
         val element = folders.find { it.id == id }
-        if(element == null) throw IllegalArgumentException("Id doesn't exist for this folder")
+        if(element == null) throw NotFoundEchoNoteException("Id doesn't exist for this folder")
         element.description = description
         element.updated_on = dateTimeCreator()
-        persistence.saveFolders(folders)
+        persistence.saveFolder(element)
         notifySubscribers()
     }
 
     suspend fun del(id: Long) {
+        if (folders.size <= 1) throw IllegalStateEchoNoteException("Cannot delete all folders")
         persistence.deleteFolder(id)
         folders.removeIf{it.id == id}
         notifySubscribers()
